수정사항 반영

+++ src/main/java/com/dsmpear/main/user_backend_v2/payload/response/MyPageReportResponse.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @Builder
 public class MyPageReportResponse {
 
-    private Integer reportId;
+    private Long reportId;
 
     private String title;
